% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_den.R
\name{observe_den}
\alias{observe_den}
\title{Observe the denominator}
\usage{
observe_den(sim, omega = 1)
}
\arguments{
\item{sim}{Simulated data from sim_data}

\item{omega}{Weighting parameter for the radius}
}
\value{
The denominator of the test statistic
}
\description{
This will return the denominator of the test statistic for data.
}
