% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_num.R
\name{observe_num}
\alias{observe_num}
\title{Observe the numerator}
\usage{
observe_num(sim, omega = 1)
}
\arguments{
\item{sim}{Simulated data from sim_data}

\item{omega}{Weighting parameter for the radius}
}
\value{
The numerator of the test statistic
}
\description{
This will return the numerator of the test statistic for data.
}
