% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate Data for IV-ANOVA}
\usage{
sim_data(group_sizes, dist, corr, params1, params2, ...)
}
\arguments{
\item{group_sizes}{Vector of group sizes}

\item{dist}{Distribution to generate data from (string)}

\item{corr}{Correlation between center and radius}

\item{params1}{First set of parameters for the distribution}

\item{params2}{Second set of parameters for the distribution}

\item{...}{Additional parameters to be passed into genCorGen}
}
\value{
A list, where each element is the data for a group
}
\description{
This function allows the user to simulate interval data with different
distributions and center/range correlations.
}
