% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_theory_den.R
\name{sim_theory_den}
\alias{sim_theory_den}
\title{Simulate denominator for limiting distribution}
\usage{
sim_theory_den(dist, params1, params2, omega = 1)
}
\arguments{
\item{dist}{Original distribution for data}

\item{params1}{Original parameters for data}

\item{params2}{Original parameters for data}

\item{omega}{Weighting parameter for radii}
}
\value{
One number, a linear combination of the variances
}
\description{
This function allows the user to get the denominator of the limiting distribution
}
