% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_theory_num.R
\name{sim_theory_num}
\alias{sim_theory_num}
\title{Simulate theoretical distribution}
\usage{
sim_theory_num(g, corr, dist, params1, params2, omega = 1, trials)
}
\arguments{
\item{g}{Number of groups}

\item{corr}{Correlation between the center and radius}

\item{dist}{Distribution of the original data, used in getting variance}

\item{params1}{First set of parameters for the data of original distribution}

\item{params2}{Second set of parameters for the data of original distribution}

\item{omega}{Weighting parameter for radii}

\item{trials}{Number of observations used to construct distribution}
}
\value{
A vector of values that simulate the limiting distribution
}
\description{
This function simulates the theoretical limiting distribution for IV-ANOVA.
It requires the distribution of the data and the original parameters.
}
